refactor: split serial_cmds.py into serial/ package
Split eab/cli/serial_cmds.py (453 lines) into modular package:

Structure:
- __init__.py (16 lines): Public API exports
- _helpers.py (84 lines): Internal helpers
- status_cmds.py (136 lines): cmd_status, cmd_tail, cmd_alerts, cmd_events
- interaction_cmds.py (195 lines): cmd_send, cmd_wait, cmd_wait_event
- capture_cmds.py (72 lines): cmd_capture_between

Test results: 1090/1103 passing (98.8%)
Serial-specific tests: 4/4 passing
