+++ eab/cli/serial/__init__.py
@@ -1,0 +1,16 @@
+"""Serial monitoring and communication commands for eabctl."""
+
+from .status_cmds import cmd_status, cmd_tail, cmd_alerts, cmd_events
+from .interaction_cmds import cmd_send, cmd_wait, cmd_wait_event
+from .capture_cmds import cmd_capture_between
+
+__all__ = [
+    "cmd_status",
+    "cmd_tail",
+    "cmd_alerts",
+    "cmd_events",
+    "cmd_send",
+    "cmd_wait",
+    "cmd_wait_event",
+    "cmd_capture_between",
+]
